app/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.27
langchain-chroma==0.2.5
langgraph==0.5.4
pypdf==5.9.0
chromadb==1.0.15
tavily-python==0.7.10
python-dotenv==1.1.1